Skill and other additiong

+++ WebApi/Services/CharecterService/CharecterService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 using WebApi.Data;
 using WebApi.DTOs.Character;
 using WebApi.Models;
@@ -8,29 +9,32 @@ namespace WebApi.Services.CharecterService
 {
     public class CharecterService : ICharecterService
     {
-        //private static List<Character> characters = new List<Character> {
-        //    new Character(),
-        //    new Character{Id=1, Name="Kumar"}
-        //};
-
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharecterService(IMapper mapper, DataContext context)
+        public CharecterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
+            .FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharecter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
             Character character = _mapper.Map<Character>(newCharecter);
+            character.User = await _context.Users.FirstOrDefaultAsync(x => x.Id == GetUserId());
+
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
 
             serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
                 .Select(c => _mapper.Map<GetCharacterDto>(c))
                 .ToListAsync();
 
@@ -43,13 +47,24 @@ namespace WebApi.Services.CharecterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(o => o.Id == id);
-                
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(o => o.Id == id && o.User.Id == GetUserId());
 
-                response.Data = await _context.Characters
-                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+
+                    response.Data = await _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found.";
+                }
+
             }
             catch (Exception ex)
             {
@@ -62,18 +77,24 @@ namespace WebApi.Services.CharecterService
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAll()
         {
-            var response=new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var response = new ServiceResponse<List<GetCharacterDto>>();
+            var dbCharacters = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .Where(o => o.User.Id == GetUserId()).ToListAsync();
 
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-            
+
             return response;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> GetById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbcharacter =await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbcharacter = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbcharacter);
             return serviceResponse;
         }
@@ -85,17 +106,71 @@ namespace WebApi.Services.CharecterService
             try
             {
                 var character = await _context.Characters
+                    .Include(o => o.User)
                     .FirstOrDefaultAsync(o => o.Id == updateCharacterDto.Id);
 
-                character.Name = updateCharacterDto.Name;
-                character.HitPoints = updateCharacterDto.HitPoints;
-                character.Strength = updateCharacterDto.Strength;
-                character.Defense = updateCharacterDto.Defense;
-                character.Intelligence = updateCharacterDto.Intelligence;
-                character.Class = updateCharacterDto.Class;
+                if (character.User.Id == GetUserId())
+                {
 
-                //_mapper.Map(updateCharacterDto, character);
+                    character.Name = updateCharacterDto.Name;
+                    character.HitPoints = updateCharacterDto.HitPoints;
+                    character.Strength = updateCharacterDto.Strength;
+                    character.Defense = updateCharacterDto.Defense;
+                    character.Intelligence = updateCharacterDto.Intelligence;
+                    character.Class = updateCharacterDto.Class;
 
+                    //_mapper.Map(updateCharacterDto, character);
+
+                    await _context.SaveChangesAsync();
+
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
+
+            }
+            catch (Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
+
+            return response;
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharecterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                var character = await _context.Characters
+                    .Include(c => c.Weapon)
+                    .Include(c => c.Skills)
+                    .FirstOrDefaultAsync(c => c.Id == newCharecterSkill.CharacterId
+                    && c.User.Id == GetUserId());
+
+                if (character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Character not found.";
+
+                    return response;
+                }
+
+                var skill = await _context.Skills.FirstOrDefaultAsync(s => s.Id == newCharecterSkill.SkillId);
+                if (skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill not found.";
+
+                    return response;
+                }
+
+                character.Skills.Add(skill);
                 await _context.SaveChangesAsync();
 
                 response.Data = _mapper.Map<GetCharacterDto>(character);
